revise add task -> make prio and category dinamic.
at prio if clicked a button then changes classes and thus the style. You can clicked the same button to toggle classes or you click on a other button at active and the on befor remove. at category you can select, add or delete categorys. you will add a category start writing in the input field and the arrow changes in a add img. you added with the add image and this changes in arrow an the input field blanks. you delete an category entry by clicked on the bin image in the dropdown.